Initial refactor in progress
