+++ container/shipit/openshift/constants.py
