+++ container/shipit/openshift/k8s_config.py
