+++ container/shipit/openshift/k8s_deployment.py
