+++ container/shipit/openshift/k8s_playbook.py
