+++ container/shipit/openshift/k8s_service.py
